Issue #179 Fix
API contained calls to logging, which would try to query the database
for configuration. If database connectivity was the reason for the need
to log, then the service or calling application would enter a loop until
it crashed or a user intervenes.

Removed many calls to logging from within the API. The calling
application should handle any logging. Surfaced exceptions back to
calling application(s) so appropriate logging messages could be
generated.