Service Debugging Framework Added a Service Debugging Framework to allow the Windows Service to be executed within Visual Studio without the need to install the service. The functionality is controlled by the use of a C# compiler constant and compiler directives. By uncomment the following line:
#define OSAESERVICECONTROLLER

from either or both the ClientService.cs file or OSAEService.cs file. The objective is to improve developer efficiency and reduce debugging time. When enabled (the OSAESERVICECONTROLLER constant being defined) the OSAEServiceController.cs form is display allowing the service to be started, paused, and/or stopped via a Windows Form UI. The service can be installed with or without the OSAESERVICECONTROLLER constant being installed.